Rename _shemaPartials to _schemaPartials
