+++ back/src/dbModels/session.ts
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose';
 
-import { simpleIdDBSchema } from './_shemaPartials';
+import { simpleIdDBSchema } from './_schemaPartials';
 import { getConfig } from '../core/config';
 
 export interface Session {
